refactor(node): move Node to node crate
