+++ core/broadcast/src/tests.rs
@@ -7,6 +7,7 @@ use lightning_application::config::ApplicationConfig;
 use lightning_interfaces::prelude::*;
 use lightning_interfaces::schema::broadcast::{Frame, Message};
 use lightning_interfaces::types::{Genesis, GenesisNode, NodePorts, Topic};
+use lightning_node::Node;
 use lightning_notifier::Notifier;
 use lightning_pool::{Config as PoolConfig, PoolProvider};
 use lightning_rep_collector::ReputationAggregator;
