+++ core/dack-aggregator/Cargo.toml
@@ -23,6 +23,7 @@ lightning-workspace-hack.workspace = true
 lightning-test-utils = { path = "../test-utils" }
 lightning-signer = { path = "../signer" }
 lightning-application = { path = "../application", features = ["test"] }
+lightning-node.workspace = true
 lightning-notifier = { path = "../notifier" }
 fleek-crypto.workspace = true
 tempfile.workspace = true
