+++ core/handshake/Cargo.toml
@@ -56,6 +56,7 @@ lightning-rpc = { path = "../rpc" }
 lightning-service-executor = { path = "../service-executor" }
 lightning-blockstore = { path = "../blockstore/" }
 lightning-test-utils = { path = "../test-utils" }
+lightning-node.workspace = true
 clap = { version = "4.4.6", features = ["derive"] }
 bincode = "1.3"
