+++ core/handshake/benches/mock.rs
@@ -9,6 +9,7 @@ use lightning_handshake::handshake::Handshake;
 use lightning_handshake::schema;
 use lightning_handshake::transports::mock::dial_mock;
 use lightning_interfaces::prelude::*;
+use lightning_node::Node;
 use lightning_service_executor::shim::ServiceExecutor;
 use lightning_service_executor::test_services::io_stress;
 use lightning_signer::Signer;
