+++ core/origin-demuxer/Cargo.toml
@@ -22,5 +22,6 @@ lightning-blockstore = { path = "../blockstore" }
 lightning-indexer = { path = "../indexer" }
 lightning-signer = { path = "../signer" }
 lightning-test-utils = { path = "../test-utils" }
+lightning-node.workspace = true
 cid.workspace = true
 tempfile.workspace = true
