+++ core/origin-demuxer/src/tests.rs
@@ -15,6 +15,7 @@ use lightning_interfaces::types::{
     NodePorts,
     OriginProvider,
 };
+use lightning_node::Node;
 use lightning_signer::Signer;
 use lightning_test_utils::consensus::{Config as ConsensusConfig, MockConsensus, MockForwarder};
 use lightning_test_utils::json_config::JsonConfigProvider;
