+++ core/origin-ipfs/Cargo.toml
@@ -33,4 +33,5 @@ lightning-blockstore = { path = "../blockstore" }
 lightning-indexer = { path = "../indexer" }
 lightning-signer = { path = "../signer" }
 lightning-test-utils = { path = "../test-utils" }
+lightning-node.workspace = true
 tempfile.workspace = true
