+++ core/test-utils/src/e2e/node.rs
@@ -13,6 +13,7 @@ use lightning_application::Application;
 use lightning_broadcast::Broadcast;
 use lightning_checkpointer::Checkpointer;
 use lightning_interfaces::prelude::*;
+use lightning_node::Node;
 use lightning_notifier::Notifier;
 use lightning_pool::PoolProvider;
 use lightning_rep_collector::MyReputationReporter;
